+++ src/app/postDetail/[postId]/interfaces.ts
@@ -1,0 +1,32 @@
+export interface IComment {
+  commentId: number;
+  userId: number;
+  content: string;
+  created: string;
+  nickname: string;
+  imageUrl: string;
+}
+
+interface ITag {
+  tagId: number;
+  tagName: string;
+}
+
+export interface IPostData {
+  postId: number;
+  imageUrl: string;
+  nickname: string;
+  title: string;
+  created: string;
+  category: string;
+  content: string;
+  voteCount: number;
+  option1: string;
+  option1Count: number;
+  option2: string;
+  option2Count: number;
+  tags: ITag[];
+  commentCount: number;
+  comments: IComment[];
+  selectedOption: string;
+}
