+++ src/app/postDetail/[postId]/postDetail.module.css
@@ -107,6 +107,7 @@
   height: 52px;
   margin-top: 20px;
   border: 1px solid var(--main-primary-600);
+  border-radius: 100px;
   background-color: #f5fffa;
   color: var(--main-primary-500);
 }
@@ -121,6 +122,7 @@
   height: 52px;
   margin-top: 8px;
   border: 1px solid var(--main-primary-600);
+  border-radius: 100px;
   background-color: #f5fffa;
   color: var(--main-primary-500);
 }
@@ -250,10 +252,12 @@
 }
 
 .moreCommentButton {
+  width: 100%;
+  height: 52px;
   font-size: 14px;
   font-weight: 500;
   line-height: 21px;
   color: var(--body-900);
-  background-color: var(--background-100);
   border: 1px solid var(--body-400);
+  border-radius: 100px;
 }
