+++ src/utils/foramattime.ts
@@ -1,0 +1,35 @@
+export function formatDay(dateString: Date | string | number) {
+  const date = new Date(dateString);
+  const month = (date.getMonth() + 1).toString().padStart(2, "0");
+  const day = date.getDate().toString().padStart(2, "0");
+  return `${month}/${day}`;
+}
+
+export function formatTime(dateString: Date | string | number) {
+  const date = new Date(dateString);
+  const hours = date.getHours().toString().padStart(2, "0");
+  const minutes = date.getMinutes().toString().padStart(2, "0");
+  return `${hours}:${minutes}`;
+}
+
+export function formatRelativeTime(dateString: Date | string | number) {
+  const date = new Date(dateString);
+  const now = new Date();
+  const diffMs = now.getTime() - date.getTime();
+  const diffMin = Math.round(diffMs / (1000 * 60));
+
+  if (diffMin < 1) {
+    return "방금 전";
+  } else if (diffMin < 60) {
+    return `${diffMin}분 전`;
+  } else if (diffMin < 24 * 60) {
+    const diffHours = Math.floor(diffMin / 60);
+    const remainingMins = diffMin % 60;
+    return `${diffHours}시간 ${remainingMins}분 전`;
+  } else {
+    const year = date.getFullYear();
+    const month = (date.getMonth() + 1).toString().padStart(2, "0");
+    const day = date.getDate().toString().padStart(2, "0");
+    return `${year}.${month}.${day}`;
+  }
+}
